Place mutate to labeled record card
