+++ asreview/webapp/src/ProjectComponents/HistoryComponents/LabeledRecordCard.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
+import { useMutation, useQueryClient } from "react-query";
 import TruncateMarkup from "react-truncate-markup";
 import {
   Card,
@@ -15,6 +16,7 @@ import { styled } from "@mui/material/styles";
 import FavoriteIcon from "@mui/icons-material/Favorite";
 import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 
+import { ProjectAPI } from "../../api/index.js";
 import { mapStateToProps } from "../../globals.js";
 
 const PREFIX = "LabeledRecordCard";
@@ -39,7 +41,46 @@ const Root = styled("div")(({ theme }) => ({
 }));
 
 const LabeledRecordCard = (props) => {
+  const queryClient = useQueryClient();
   const [recordReadMore, setRecordReadMore] = React.useState(null);
+
+  const { mutate } = useMutation(ProjectAPI.mutateClassification, {
+    onSuccess: (data, variables) => {
+      // update cached data
+      queryClient.setQueryData(
+        [
+          props.label === "relevant"
+            ? "fetchRelevantLabeledRecord"
+            : "fetchIrrelevantLabeledRecord",
+          {
+            project_id: props.project_id,
+            select: props.label === "relevant" ? "included" : "excluded",
+          },
+        ],
+        (prev) => {
+          return {
+            ...prev,
+            pages: prev.pages.map((page) => {
+              return {
+                ...page,
+                result: page.result.map((value) => {
+                  return {
+                    ...value,
+                    included:
+                      value.id === variables.doc_id
+                        ? value.included === 1
+                          ? 0
+                          : 1
+                        : value.included,
+                  };
+                }),
+              };
+            }),
+          };
+        }
+      );
+    },
+  });
 
   return (
     <Root>
@@ -82,7 +123,7 @@ const LabeledRecordCard = (props) => {
               <IconButton
                 className={classes.icon}
                 onClick={() => {
-                  props.mutateClassification({
+                  mutate({
                     project_id: props.project_id,
                     doc_id: value.id,
                     label: value.included,
